re-arrange vault struct; break out terraform ops into file
